Move theme.ts to project for workaround until primevue 4
