+++ arches_references/src/arches-references/components/editor/ItemType.vue
@@ -5,7 +5,7 @@ import { useGettext } from "vue3-gettext";
 import InputSwitch from "primevue/inputswitch";
 import { useToast } from "primevue/usetoast";
 
-import { ARCHES_CHROME_BLUE } from "@/arches/theme.ts";
+import { ARCHES_CHROME_BLUE } from "@/arches-references/theme";
 import { patchItem } from "@/arches-references/api.ts";
 import {
     DEFAULT_ERROR_TOAST_LIFE,
